Add back support for the Carbon interface (Mac::File and friends) under Mac OS 10.3 and below, where the API used by Mac::SystemDirectory appears not to be present.
The change adds back the Carbon interface _provided_ we are not a 64-bit
Perl and the most significant part of the uname OS version is less than
8 (this is the Darwin version corresponding to Mac OS 10.4 (Tiger)).

The patch has been tried on 32-bit Perls under Mac OS 10.3, 10.4, and
10.5. It will specify requirements that fail to build under a 64-bit
Perl under Mac OS 10.3 and below, but so do the unpatched
File::HomeDir 0.93 and File::HomeDir 0.91.

You are urged to sanity-check the patch with any Mac OS X gurus you have
on tap.